requirements.txt
PySide6>=6.7
PySide6-Addons>=6.7
PySide6-Essentials>=6.7
numpy>=1.24
pandas>=2.0
Pillow>=10.0,<11
matplotlib>=3.8
requests>=2.31
opencv-python>=4.9
albumentations>=1.4
scikit-learn>=1.4
pydantic>=2.6
reportlab>=4.1
fastapi>=0.111
uvicorn>=0.30
torch>=2.2
torchvision>=0.17
gradio>=4.0
phe>=1.5
pyinstaller>=6.6
pyinstaller-hooks-contrib>=2024.6
